Incremental commit.
Implment session controller.
Implment web site.

+++ src/main/java/cn/com/xinli/portal/rest/RestSessionService.java
@@ -1,11 +1,13 @@
 package cn.com.xinli.portal.rest;
 
-import cn.com.xinli.portal.PortalException;
 import cn.com.xinli.portal.Session;
+import cn.com.xinli.portal.SessionNotFoundException;
 import cn.com.xinli.portal.SessionService;
 import cn.com.xinli.portal.persist.SessionEntity;
 import cn.com.xinli.portal.persist.SessionRepository;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
 
 import java.util.Calendar;
+import java.util.Date;
 
 /**
  * Project: portal
@@ -21,6 +24,8 @@ import java.util.Calendar;
  */
 @Service
 public class RestSessionService implements SessionService, InitializingBean {
+    /** Log. */
+    private static final Log log = LogFactory.getLog(RestSessionService.class);
 
     @Autowired
     private SessionRepository sessionRepository;
@@ -31,28 +36,56 @@ public class RestSessionService implements SessionService, InitializingBean {
     }
 
     @Override
-    @Transactional
-    public Session createSession(String ip, String mac, String nasId) throws PortalException {
-        SessionEntity session = new SessionEntity();
-        session.setIp(ip);
-        session.setMac(mac);
-        session.setNasId(nasId);
-        session.setStartDate(Calendar.getInstance().getTime());
-        session.setUsername(StringUtils.join(ip, " ", mac));
+    public Session createSession(Session session) {
+        SessionEntity existed = sessionRepository.find(Session.pair(session.getIp(), session.getMac()));
+        if (existed != null) {
+            log.warn("> session already exists." + existed);
+            return existed;
+        }
 
-        return sessionRepository.save(session);
+        SessionEntity entity = (SessionEntity) session;
+        return sessionRepository.save(entity);
     }
 
     @Override
     @Transactional
-    public Session getSession(long id) throws PortalException {
-        return sessionRepository.findOne(id);
+    public Session getSession(long id) throws SessionNotFoundException {
+        Session found = sessionRepository.findOne(id);
+        if (found == null) {
+            throw new SessionNotFoundException(id);
+        }
+        return found;
     }
 
     @Override
     @Transactional
-    public boolean removeSession(long id) throws PortalException {
+    public void removeSession(long id) throws SessionNotFoundException {
+        if (sessionRepository.findOne(id) == null) {
+            throw new SessionNotFoundException(id);
+        }
         sessionRepository.delete(id);
-        return true;
+    }
+
+    @Override
+    @Transactional
+    public Session find(String ip, String mac) {
+       return sessionRepository.find(Session.pair(ip, mac));
+    }
+
+    @Override
+    @Transactional
+    public Session update(long id, long timestamp) {
+        SessionEntity found = sessionRepository.findOne(id);
+        if (found != null) {
+            throw new SessionNotFoundException(id);
+        }
+
+        Calendar calendar = Calendar.getInstance();
+        calendar.setTimeInMillis(timestamp);
+        Date date = calendar.getTime();
+
+        found.setLastModified(date);
+
+        return sessionRepository.save(found);
     }
 }
